Pre-release identifiers

A class representing a collection of identifiers, which are used for representing pre-release versions.

There are two functions to create the pre_release_ids vector:

pre_release_ids() is a low-level constructor for creating pre-release identifiers from individual components.

parse_pre_release_ids() parses a character vector into pre-release identifiers.

Usage

new_pre_release_ids(id1 = character(), id2 = "", id3 = "", id4 = "", id5 = "")

parse_pre_release_ids(x)

Arguments

id1, id2, id3, id4, id5

Single pre-release identifiers. Each identifier can be something to be cast to a pre_release_identifier vector by vctrs::vec_cast(). All components must be of the same length or length 1 (will be recycled).

x

A character vector representing pre-release identifiers. Each identifier separated by a dot (.) will be parsed as a pre_release_identifier.

Value

A pre_release_ids vector.

Details

Internally, pre_release_ids store up to 5 pre_release_identifier. So this can't represent pre-release identifiers with more than 5 components. If passing character containing more than 5 components to parse_pre_release_ids(), it will throw an error.

If the components are empty, they are treated as the highest precedence pre-release ids, which is used to indicate that the version is not a pre-release version.

Examples

# Each components are concatenated with a dot
new_pre_release_ids("rc", 1:3)
#> rc.1
#> rc.2
#> rc.3

ids <- parse_pre_release_ids(
  c("", "alpha.beta", "alpha.1", "beta", "beta.11", "beta.2")
)
ids
#> <empty>
#> alpha.beta
#> alpha.1
#> beta
#> beta.11
#> beta.2

# Empty ids have the highest precedence
# (Used to indicate not a pre-release version)
vctrs::vec_sort(ids)
#> alpha.1
#> alpha.beta
#> beta
#> beta.2
#> beta.11
#> <empty>

# Can be compared with string notation
ids[ids > "beta.2"]
#> <empty>
#> beta.11

# Limitation: only support up to 5 components
parse_pre_release_ids("a.b.c.d.e")
#> a.b.c.d.e
try(parse_pre_release_ids("a.b.c.d.e.f"))
#> Error in parse_pre_release_ids("a.b.c.d.e.f") : 
#>   Unsupported pre-release identifiers in 'a.b.c.d.e.f'.
#> ! Only up to 5 pre-release identifiers are supported, got 6.